rd: Add qrtr nameserver
A qrtr nameserver implementation is required for Qualcomm Hexagon DSP
services to work properly. Power supply interfaces (fuel gauge and
charger) are provided by the aDSP on the Snapdragon 888 and newer
platforms, so we need to boot it and make it work before we can get
power measurements.

This was compiled from https://github.com/andersson/qrtr.