init: Boot and initialize aDSP on Snapdragon 888
Power supply interfaces (fuel gauge and charger) are provided by the
aDSP on the Snapdragon 888 and newer platforms, so we need to boot it
and make it work before we can get power measurements.

We check for these newer platforms and try to boot the aDSP
automatically if the kernel version is 5.x and a Qualcomm CPU is
reported in cpuinfo.