Change base_power calculation to prevent too low values
The base_power is currently calculated as the minimum value of the power measurements. This value can sometimes be lower than the actual average base power draw, resulting in too low base_power and thus too high per-frequency power and too low efficiency results. This can be seen on some results already, especially the proposed results of the SM7325.
This patch changes the calculation of the base_power from being the smallest measurement (min) to the average measurement (median).